+++ frontend/ui/src/app/interfaces/abstraction/leader.ts
@@ -2,5 +2,5 @@ import {Human} from "./human";
 import {Battle} from "../models/politics/military/battle";
 
 export interface Leader extends Human {
-  battles: Battle[];
+  battles?: Battle[];
 }